fix dark-mode, scroll, remove extra files
