+++ src/App.js
@@ -3,7 +3,7 @@ import {
   Routes,
   Route,
 } from "react-router-dom";
-import Navbar2 from './components/navBar/Navbar2';
+import NavBar from './components/navBar/Navbar';
 import Home from './pages/home/Home';
 import Shop from './pages/shop/Shop';
 import CreateComic from './pages/create-comic/CreateComic';
@@ -14,7 +14,7 @@ function App() {
   return (
     <div className="App">
       <Routes>
-        <Route path="/" element={<LandingPage />} />
+        <Route exact path="/" element={<LandingPage />} />
         <Route path="/home" element={<Home />} />
         <Route path="/shop" element={<Shop />} />
         <Route path={"/cardDetail/:id"} element={<CardDetail />} />
