+++ src/redux/reducers/index.js
@@ -4,7 +4,10 @@ const initialState = {
   comic: {},
   issue: {},
   cart_shopping: 0,
-  theme: "light"
+  theme_params: {
+    theme: "light",
+    state: false
+  }
 };
 
 
@@ -41,7 +44,7 @@ const rootReducer = (state = initialState, action) => {
     case "SET_THEME":
       return {
         ...state,
-        theme: action.payload,
+        theme_params: action.payload,
       }
